+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
-- [Fixed NotificationsManager hanging toasts](https://github.com/multiversx/mx-sdk-dapp/pull/1622)
+- [Added NotificationsFeedManager basic tests](https://github.com/multiversx/mx-sdk-dapp/pull/1625)
+- [Fixed NotificationsFeedManager hanging toasts](https://github.com/multiversx/mx-sdk-dapp/pull/1622)
 - [Added basic ToastManager test](https://github.com/multiversx/mx-sdk-dapp/pull/1619)
 
 ## [[5.2.9](https://github.com/multiversx/mx-sdk-dapp/pull/1621)] - 2025-10-15